tests/wpt/web-platform-tests/css/css-text/white-space/text-wrap-balance-001.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#text-wrap">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
#container {
  width: 20ch;
}
.balance {
  text-wrap: balance;
}
</style>
<div id="container"></div>
<script>
const container = document.getElementById('container');
for (const text of [
    'Balancing should',
    'Balancing should not change',
    'Balancing should not change the number of lines.',
  ]) {
  const normal = document.createElement('div');
  const balance = document.createElement('div');
  normal.textContent = text;
  balance.textContent = text;
  balance.classList.add('balance');
  container.appendChild(normal);
  container.appendChild(balance);
  test(() => {
    assert_equals(normal.offsetHeight, balance.offsetHeight);
  });
}
</script>
